+++ sw360/__init__.py
@@ -11,10 +11,12 @@ __version__ = (1, 11, 2)
 
 from .sw360_api import SW360
 from .sw360error import SW360Error
+from .sw360keycloak import SW360Keycloak
 from .sw360oauth2 import SW360OAuth2
 
 __all__ = [
     "SW360",
     "SW360Error",
-    "SW360OAuth2"
+    "SW360OAuth2",
+    "SW360Keycloak"
 ]